Fix: downgrade to influxDB 2
We primaraly used  influxdb 3 Core but it show's quickly a real show stopper. This TSDB is mainly built to work on recent time series with decent performance on time range from now until 72h. Above this time range, due to data fragmentation, measurement can be  spread across a lot of parquets files and above 492 files request are limited by the db engine. In our case this is a no go as we need to shows much larger time range.
The compactor is the tool that can re-arrange parquet file to avoid this issue, however it is available only on the entreprise version of influxdb 3. Details can be read here : https://www.influxdata.com/blog/influxdb3-open-source-public-alpha-jan-27/

This commit downgrade influxdb version to 2.